requirements.txt
certifi==2019.3.9
chardet==3.0.4
cloudinary==1.21.1
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-autoslug-iplweb==1.9.4.dev0
django-cities-light==3.5.0
django-cors-middleware==1.3.1
django-rest-swagger==2.2.0
djangorestframework==3.9.4
django-filter
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
progressbar2==3.42.0
python-utils==2.3.0
pytz==2019.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
Unidecode==1.1.0
uritemplate==3.0.0
urllib3==1.25.3
Pillow
django-storages
django-model-utils==3.2.0
djangorestframework-recursive==0.1.2
whitenoise==4.1.3
dj-database-url==0.5.0
gunicorn==19.9.0
